fix
コンポーネントの構成を変更した
