+++ front/src/index.tsx
@@ -1,37 +1,16 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import { BrowserRouter } from 'react-router-dom';
 import './index.css';
 import reportWebVitals from './reportWebVitals';
-import Box from '@mui/material/Box';
-import CssBaseline from '@mui/material/CssBaseline';
-
-import Sidebar from './Sidebar';
-import Login from './Login';
-import Home from './Home';
-import TweetDetail from './TweetDetail';
-import ProfilePage from './ProfilePage';
-import SettingsProfile from './SettingsProfile';
+import Main from './Main';
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 root.render(
     <React.StrictMode>
-        <Box sx={{ display: 'flex' }}>
-            <BrowserRouter>
-                <CssBaseline />
-                <Sidebar />
-                {/* /loginでは表示しない */}
-                <Box component="main" sx={{ flexGrow: 1, bgcolor: 'background.default', p: 3 }}>
-                    <Routes>
-                        <Route path="login" element={<Login />} />
-                        <Route path="/" element={<Home />} />
-                        <Route path="tweet/:tweetId" element={<TweetDetail />} />
-                        <Route path=":username" element={<ProfilePage />} />
-                        <Route path="settings/:username" element={<SettingsProfile />} />
-                    </Routes>
-                </Box>
-            </BrowserRouter>
-        </Box>
+        <BrowserRouter>
+            <Main />
+        </BrowserRouter>
     </React.StrictMode>
 );
